moved target annotation and parsing to separate module
